+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version numbers are roughly based on [Semantic Versioning](https://semver.org/sp
 
 * Seneschal details on the Edit Group and Report pages now taken from the `warrants`
   (Regnumator) table.
+* Tweaked format of Announce event notice subject - date before event name.
 
 ## [2.14.0] - 2023-02-09